pyproject.toml
[project]
name = "stockelper-llm"
version = "0.1.0"
description = "AI-powered stock investment assistant using LangGraph multi-agent system"
# readme = "README.md"  # Optional for Docker builds
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.111.0",
    "requests",
    "httpx",
    "aiohttp",
    "numpy",
    "openai",
    "uvicorn",
    "finance-datareader",
    "opendartreader",
    "plotly",
    "statsmodels",
    "asyncpg",
    "psycopg[binary,pool]==3.2.*",
    "greenlet",
    "sqlalchemy>=2.0",
    "langchain>=1.0.0",
    "langchain-classic>=1.0.0",
    "langgraph",
    "langfuse",
    "langgraph-checkpoint-postgres",
    "langchain-openai",
    "langchain-community",
    "langchain-neo4j",
    "neo4j",
    "pymongo",
    "motor",
    "youtube-transcript-api",
    "google-api-python-client",
    "prophet",
    "pandas",
    "matplotlib",
    "mojito2",
    "python-dotenv",
]

[project.optional-dependencies]
ml = [
    "langchain>=1.0.0",
    "langchain-classic>=1.0.0",
    "langgraph",
    "langchain-community", 
    "langchain-openai",
    "langchain-huggingface",
    "langfuse",
    "pymongo",
    "lancedb",
    "FlagEmbedding",
    "prophet",
    "motor",
    "youtube-transcript-api",
    "google-api-python-client",
    "langgraph-checkpoint-postgres",
    "langchain-neo4j",
]
dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "isort",
    "flake8",
    "mypy",
]

# Optional groups to avoid heavy deps by default
ui = [
    "streamlit",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stockelper_llm"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
